:truck: Rename: teamBuilding 오타 수정 (#98)
